jimmied up install with multiple actions

+++ install.sh
@@ -1,37 +1,47 @@
 #! /bin/bash
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-echo 'Working directory: $DIR'
 
-# bash
-echo 'Installing Bash config'
-rm ~/.bash* 2>/dev/null
-ln -s $DIR/bash/rc ~/.bashrc
-ln -s $DIR/bash/profile ~/.bash_profile
-ln -s $DIR/bash/include ~/.bash_include
-ln -s $DIR/bash/aliases ~/.bash_aliases
-ln -s $DIR/bash/functions ~/.bash_functions
-ln -s $DIR/bash/logout ~/.bash_logout
+#
+# default action: install environment
+#
+if [[ -z "$1" ]]; then
+    echo "Working directory: ${DIR}"
 
-# vim
-echo 'Insalling Vim config'
-rm ~/.vim* 2>/dev/null
-ln -s $DIR/vim/rc ~/.vimrc
-#ln -s $DIR/vim/vim/ ~/test/.vim
+    # bash
+    echo 'Installing Bash config'
+    rm ~/.bash* 2>/dev/null
+    ln -s $DIR/bash/rc ~/.bashrc
+    ln -s $DIR/bash/profile ~/.bash_profile
+    ln -s $DIR/bash/include ~/.bash_include
+    ln -s $DIR/bash/aliases ~/.bash_aliases
+    ln -s $DIR/bash/functions ~/.bash_functions
+    ln -s $DIR/bash/logout ~/.bash_logout
 
-# git
-echo 'Installing Git config'
-rm ~/.git* 2>/dev/null
-ln -s $DIR/git/config ~/.gitconfig
-ln -s $DIR/git/completion.bash ~/.git-completion.bash
+    # vim
+    echo 'Installing Vim config'
+    rm ~/.vim* 2>/dev/null
+    ln -s $DIR/vim/rc ~/.vimrc
+    #ln -s $DIR/vim/vim/ ~/test/.vim
 
-# gem
-echo 'Installing Gem config'
-rm ~/.gem* 2>/dev/null
-ln -s $DIR/gem/rc ~/.gemrc
+    # git
+    echo 'Installing Git config'
+    rm ~/.git* 2>/dev/null
+    ln -s $DIR/git/config ~/.gitconfig
+    ln -s $DIR/git/ignore ~/.gitignore
+    ln -s $DIR/git/completion.bash ~/.git-completion.bash
 
-# pow
-echo 'Installing Pow config'
-rm ~/.pow* 2>/dev/null
-ln -s $DIR/pow/config ~/.powconfig
+    # gem
+    echo 'Installing Gem config'
+    rm ~/.gem* 2>/dev/null
+    ln -s $DIR/gem/rc ~/.gemrc
 
+    # pow
+    echo 'Installing Pow config'
+    rm ~/.pow* 2>/dev/null
+    ln -s $DIR/pow/config ~/.powconfig
+elif [[ "$1" == "header" ]]; then
+    echo $1
+else
+    echo "Dont know what ${1} means"
+fi
